+++ docs/changes.rst
@@ -3,6 +3,8 @@ Weblate 4.14
 
 Not yet released.
 
+* Track add-on changes in a history.
+
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/84?closed=1>`__.
 
 Weblate 4.13.1